Refactor PartitionTest
